# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Gossip Girl Blog API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface AuthStatus {
  loggedIn: boolean;
}

export interface Post {
  id: number;
  text: string;
  /** @nullable */
  photoUrl?: string | null;
  /** Unix timestamp in milliseconds */
  timestamp: number;
}

export interface PostInput {
  /** @minLength 1 */
  text: string;
  /** @nullable */
  photoUrl?: string | null;
}

export interface DeleteResult {
  success: boolean;
}