# lib/api-zod/src/generated/types/loginInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Gossip Girl Blog API
 * OpenAPI spec version: 0.1.0
 */

export interface LoginInput {
  username: string;
  password: string;
}