# lib/api-zod/src/generated/types/post.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Gossip Girl Blog API
 * OpenAPI spec version: 0.1.0
 */

export interface Post {
  id: number;
  text: string;
  /** @nullable */
  photoUrl?: string | null;
  /** Unix timestamp in milliseconds */
  timestamp: number;
}