# lib/api-zod/src/generated/types/postInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Gossip Girl Blog API
 * OpenAPI spec version: 0.1.0
 */

export interface PostInput {
  /** @minLength 1 */
  text: string;
  /** @nullable */
  photoUrl?: string | null;
}