# lib/db/src/schema/posts.ts
import { pgTable, serial, text, bigint } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const postsTable = pgTable("posts", {
  id: serial("id").primaryKey(),
  text: text("text").notNull(),
  photoUrl: text("photo_url"),
  timestamp: bigint("timestamp", { mode: "number" }).notNull(),
});

export const insertPostSchema = createInsertSchema(postsTable).omit({ id: true });
export type InsertPost = z.infer<typeof insertPostSchema>;
export type Post = typeof postsTable.$inferSelect;